pyproject.toml
[tool.poetry]
name = "auto-ppt-maker"
version = "0.1.0"
description = ""
authors = ["andyzhou <andyzhou1982@163.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.2.15"
langchain-openai = "^0.1.23"
langchain-community = "^0.2.14"
python-dotenv = "^1.0.1"
pillow = "^10.4.0"
py-xid = "^1.0.0"
diffusers = "^0.30.1"
torch = "^2.4.0"
transformers = {extras = ["sentencepiece"], version = "^4.44.2"}
accelerate = "^0.33.0"
gradio = "^4.42.0"
pyunsplash = "^1.0.0rc2"
optimum-quanto = "^0.2.4"
python-pptx = "^1.0.2"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
